refactor(products): restructure state management architecture
- implement new productStore with improved state management
- replace productUseCase with context-based architecture
- add useProductStore hook for better state access
- update product context to use new store structure
