+++ packages/react/src/entities/products/hooks/useProductStore.ts
@@ -1,12 +1,24 @@
-import { createStore, useStore } from "@hanghae-plus/lib";
-import { initialProductState, productStore, type ProductStoreState } from "../productStore";
+import { createStore, type StringRecord } from "@hanghae-plus/lib";
 import { createContext, useContext } from "react";
+import { initialProductState } from "../productStore";
 
-export type ProductStore = ReturnType<typeof createStore<typeof initialProductState, unknown>>;
+export type ProductStore = {
+  store: ReturnType<typeof createStore<typeof initialProductState, unknown>>;
+  state: typeof initialProductState;
+  action: {
+    loadProducts: (resetList?: boolean) => Promise<void>;
+    loadProductsAndCategories: () => Promise<void>;
+    searchProducts: (search: string) => void;
+    setCategory: (category: StringRecord) => void;
+    setSort: (sort: string) => void;
+    setLimit: (limit: number) => void;
+    loadProductDetailForPage: (productId: string) => void;
+    loadRelatedProducts: (category2: string, excludeProductId: string) => void;
+    loadNextProducts: () => void;
+  };
+};
 
 export const ProductStoreContext = createContext<ProductStore | null>(null);
-
-export const useProductStore = (snapshot: ProductStoreState) => useStore(productStore, (state) => state, snapshot);
 
 export const useProductStoreContext = () => {
   const productStore = useContext(ProductStoreContext);
