+++ packages/react/src/entities/products/index.ts
@@ -2,4 +2,3 @@ export * from "./productStore";
 export * from "./hooks";
 export * from "./components";
 export * from "./types";
-export * from "./productUseCase";
