+++ packages/react/src/entities/products/productStore.ts
@@ -142,11 +142,6 @@ const productReducer = (state: typeof initialProductState, action: any) => {
   }
 };
 
-/**
- * 상품 스토어 생성
- */
-export const productStore = createStore(productReducer, initialProductState);
-
 export const createProductStore = (initState: Partial<ProductStoreState>) => {
   return createStore(productReducer, { ...initialProductState, ...initState });
 };
